🎉 Start flowlab and flowlab-lib
Based on offline tests, this project can become quite big. So, it will be separated into different crates: the lib with the basic implementations and the binary with the user interface (it may have two variants in the future, the CLI and GUI versions). This commit also adds the basic Device implementation as well as the TCP-specific protocol implementation. The TCP implementation is only available if the "tcp" feature is enabled.